Cleaned up waifu commands
