+++ NadekoBot.Core/Modules/Gambling/Common/Waifu/AffinityTitle.cs
@@ -1,0 +1,16 @@
+﻿namespace NadekoBot.Core.Modules.Gambling.Common.Waifu
+{
+    public enum AffinityTitle
+    {
+        Pure,
+        Faithful,
+        Defiled,
+        Cheater,
+        Tainted,
+        Corrupted,
+        Lewd,
+        Sloot,
+        Depraved,
+        Harlot
+    }
+}
