+++ NadekoBot.Core/Modules/Gambling/Common/Waifu/DivorceResult.cs
@@ -1,0 +1,10 @@
+﻿namespace NadekoBot.Core.Modules.Gambling.Common.Waifu
+{
+    public enum DivorceResult
+    {
+        Success,
+        SucessWithPenalty,
+        NotYourWife,
+        Cooldown
+    }
+}
