+++ NadekoBot.Core/Modules/Gambling/Common/Waifu/WaifuClaimResult.cs
@@ -1,0 +1,9 @@
+﻿namespace NadekoBot.Core.Modules.Gambling.Common.Waifu
+{
+    public enum WaifuClaimResult
+    {
+        Success,
+        NotEnoughFunds,
+        InsufficientAmount
+    }
+}
