+++ NadekoBot.Core/Modules/Gambling/Common/Waifu/WaifuProfileTitle.cs
@@ -1,0 +1,14 @@
+﻿namespace NadekoBot.Core.Modules.Gambling.Common.Waifu
+{
+    public struct WaifuProfileTitle
+    {
+        public int Count { get; }
+        public string Title { get; }
+
+        public WaifuProfileTitle(int count, string title)
+        {
+            Count = count;
+            Title = title;
+        }
+    }
+}
